refactor: word item 이름 변경
