+++ src/components/molecules/WordItem/WordItem.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { WordItem } from '@/components/molecules/WordDetail/WordDetail';
+import { WordItem } from '@/components/molecules/WordItem/WordItem';
 
 const meta: Meta<typeof WordItem> = {
   title: 'Molecules/WordDetail',
